refactor: explicit-width integer types across the codebase
Completes the sweep the reflectgen/GpuTelemetry commit started. Bare int,
unsigned, and long long become int32_t/uint32_t/int64_t at every site that is
genuinely ours, per the engine's explicit-width rule. Files containing a bare
integer type drop from 75 to 40.

The substitutions are ABI-identical on every target platform (int is int32_t,
unsigned is uint32_t), so this is a spelling change with no behavioural effect
— which the warning counts confirm: -Wsign-conversion is unchanged at 28 and
every other category matches the pre-sweep build exactly. (Raw warning totals
differ only because generated reflection sources fall in or out of a given
build's recompile scope; a like-for-like comparison against the previous full
build shows the same 13 sites before and after.)

Left alone deliberately, because the type is a contract rather than a choice:
- GpuTelemetry.cpp in full — it mirrors NVML's C ABI, where the layout is the
  point. Zero diff.
- stb_image / libwebp / glslang buffer and parameter types in Texture.cpp.
- GLFW callback signatures and the locals whose addresses GLFW writes through
  (Window module), and GLFW-derived key/button enum underlying types.
- Jolt overrides, notably GetNumBroadPhaseLayers returning JPH::uint.
- ImGui out-parameters — Combo/InputInt/RadioButton take int*, so retyping the
  variable behind the address would break the call.
- int main(int argc, char**), and the static_cast<int> feeding printf's %.*s
  precision argument, both mandated by the standard.
- Win32/POSIX interop in Platform.cpp.
- char used as text; only char-as-small-number became a fixed-width type.

One seam needed fixing by hand: Application::OnResize became int32_t while
SandboxApp's override was left as int precisely because the base said int, so
the two spellings disagreed (identical types, so it still compiled). Both are
int32_t now, along with the doc comment.

Validated: gcc-debug builds clean and 8/8 suites pass; ASan and TSan each build
the full tree and pass 8/8 with zero reports.

+++ apps/sandbox/src/SandboxApp.hpp
@@ -84,7 +84,7 @@ class SandboxApp : public Assisi::App::Application
     void OnUpdate(float dt) override;
     void OnRender(Assisi::Render::RenderFrame &frame) override;
     void OnImGui() override;
-    void OnResize(int width, int height) override;
+    void OnResize(int32_t width, int32_t height) override;
     void OnRenderTargetsChanged(const nvrhi::FramebufferInfo &framebufferInfo) override;
     void FlushDeferred() override;
 
@@ -416,7 +416,7 @@ class SandboxApp : public Assisi::App::Application
     // A history jump requested by clicking a History-panel row: negative = undo N
     // steps, positive = redo N. Applied at the top of the next OnUpdate (never
     // mid-ImGui, which would invalidate cached component pointers).
-    int _pendingHistorySteps = 0;
+    int32_t _pendingHistorySteps = 0;
 
     // Per-component delete confirmation: the inspector's X button arms a two-step
     // confirm for one component at a time. Scoped to an entity so switching
@@ -473,7 +473,7 @@ class SandboxApp : public Assisi::App::Application
     char _addComponentBuf[64] = {};
     // Keyboard highlight into the suggestion list: Tab/Down advance it, Up retreats,
     // editing the text resets it to the first row, Enter adds the highlighted one.
-    int _addComponentSelected = 0;
+    int32_t _addComponentSelected = 0;
 
     // NVIDIA GPU telemetry (clocks/power/util/temp) for the options overlay.
     // Lazily initialises NVML on first poll, so it costs nothing until the
@@ -485,13 +485,13 @@ class SandboxApp : public Assisi::App::Application
     // so the buffers span ~30s regardless of frame rate. _gpuTelemetryOffset is
     // the next write slot / chronological start (ImPlot Offset), _gpuTelemetryCount
     // saturates at the capacity. Only advance while the overlay is open.
-    static constexpr int                        kGpuHistory = 150; // ~30s at 5Hz
+    static constexpr int32_t                    kGpuHistory = 150; // ~30s at 5Hz
     std::array<float, kGpuHistory>              _gpuClockHistory{};
     std::array<float, kGpuHistory>              _gpuUtilHistory{};
     std::array<float, kGpuHistory>              _gpuPowerHistory{};
-    int                                         _gpuTelemetryOffset = 0;
-    int                                         _gpuTelemetryCount  = 0;
-    unsigned long long                          _lastGpuSequence    = 0;
+    int32_t                                     _gpuTelemetryOffset = 0;
+    int32_t                                     _gpuTelemetryCount  = 0;
+    uint64_t                                    _lastGpuSequence    = 0;
 
     // Eyedropper: while armed, the next scene entity-pick is written into the
     // captured EntityRef field instead of changing the selection. The target is
@@ -531,7 +531,7 @@ class SandboxApp : public Assisi::App::Application
     Assisi::Render::AssetCache _thumbnailCache;
 
     std::vector<std::string> _levelFiles;
-    int                      _selectedLevel = 0;
+    int32_t                  _selectedLevel = 0;
     char                     _saveAsName[128] = {};
     std::string              _startupLevel; // level stem to open in OnStart (CLI); empty = none.
     // A level load requested from the UI (OnImGui), applied at the next OnUpdate —
